# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Books API",
        "description": "API REST de gestion de livres avec authentification Sanctum.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Liste paginée des livres",
                "description": "Retourne la liste des livres paginée (2 par page).",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptJsonHeader"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/books?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/books?page=2"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/books?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Créer un livre",
                "description": "Crée un nouveau livre. Authentification requise.",
                "operationId": "b10f5fe35eefa47e974dd4b333d870b1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptJsonHeader"
                    },
                    {
                        "$ref": "#/components/parameters/ContentTypeJsonHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "summary",
                                    "isbn"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Fondation",
                                        "maxLength": 255,
                                        "minLength": 3
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Isaac Asimov",
                                        "maxLength": 100,
                                        "minLength": 3
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Cycle de science-fiction sur la chute et la renaissance d'un empire galactique.",
                                        "maxLength": 500,
                                        "minLength": 10
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "9780553293357",
                                        "maxLength": 13,
                                        "minLength": 13
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livre créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The isbn field must be 13 characters."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "isbn": [
                                                    "The isbn field must be 13 characters."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "1984"
                    },
                    "author": {
                        "description": "Renvoyé en majuscules",
                        "type": "string",
                        "example": "GEORGE ORWELL"
                    },
                    "summary": {
                        "type": "string",
                        "example": "Roman dystopique culte..."
                    },
                    "isbn": {
                        "type": "string",
                        "example": "9780451524935"
                    },
                    "_links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/v1/books/1"
                            },
                            "update": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/v1/books/1"
                            },
                            "delete": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/v1/books/1"
                            },
                            "all": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/v1/books"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "AcceptJsonHeader": {
                "name": "Accept",
                "in": "header",
                "description": "Force Laravel à renvoyer du JSON",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "application/json"
                }
            },
            "ContentTypeJsonHeader": {
                "name": "Content-Type",
                "in": "header",
                "description": "Type de contenu envoyé",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "application/json"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}